backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
mysql-connector-python
requests
beautifulsoup4
python-dotenv
PyJWT
pdfplumber
PyPDF2
camelot-py[cv]
pandas
aiohttp
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pymongo==4.6.0
